+++ src/main.rs
@@ -70,6 +70,10 @@ struct Options {
     #[arg(long = "initial-plugin", value_name = "PLUGIN", help_heading = None)]
     initial_plugin: Option<String>,
 
+    /// Which engine to use for OCR. Available: "tesseract", "manga-ocr". Default: "tesseract"
+    #[arg(long = "ocr-engine", value_name = "ENGINE", help_heading = None)]
+    ocr_engine: Option<String>,
+
     /// Try to open the window at the mouse cursor. Unlikely to work on wayland
     #[arg(short = 'm', long = "at-mouse", help_heading = None)]
     open_at_cursor: bool,
@@ -141,8 +145,17 @@ fn main() -> ExitCode {
         font: cli.options.font.unwrap_or(String::from("Noto Sans CJK JP")),
     };
 
+    let mut initial_ocr_model: usize = 0;
+    if let Some(ocr_engine) = cli.options.ocr_engine {
+        if ocr_engine == "tesseract" {
+            initial_ocr_model = 0;
+        } else if ocr_engine == "manga-ocr" {
+            initial_ocr_model = 1;
+        }
+    }
+
     let paused = Arc::new(AtomicBool::new(false));
-    let ocr_model = Arc::new(AtomicUsize::new(0)); // 0=Tesseract; 1=MangaOCR
+    let ocr_model = Arc::new(AtomicUsize::new(initial_ocr_model)); // 0=Tesseract; 1=MangaOCR
     let mut manga_ocr = None;
 
     #[cfg(target_os = "linux")]
@@ -181,7 +194,9 @@ fn main() -> ExitCode {
         } else if let Some(ocr_path) = cli.modes.ocr {
             match get_image_for_ocr(ocr_path) {
                 Ok(image) => {
-                    if let Err(e) = popup_dictionary::ocr(image, config, 0, &mut manga_ocr) {
+                    if let Err(e) =
+                        popup_dictionary::ocr(image, config, initial_ocr_model, &mut manga_ocr)
+                    {
                         tracing::error!("Failed while running ocr mode due to error: {e}");
                         return ExitCode::FAILURE;
                     }
@@ -233,7 +248,9 @@ fn main() -> ExitCode {
         } else if let Some(ocr_path) = cli.modes.ocr {
             match get_image_for_ocr(ocr_path) {
                 Ok(image) => {
-                    if let Err(e) = popup_dictionary::ocr(image, config, 0, &mut manga_ocr) {
+                    if let Err(e) =
+                        popup_dictionary::ocr(image, config, initial_ocr_model, &mut manga_ocr)
+                    {
                         tracing::error!("Failed while running ocr mode due to error: {e}");
                         return ExitCode::FAILURE;
                     }